Restructure TestFoundation directory
This moves the `TestFoundation` directory to a `Tests` directory to home
all the testing infrastructure.  The Foundation tests are now present
under `Tests/Foundation`.  The Test cases themselves are in the `Tests`
directory under `Tests/Foundation` and the supporting files in the
directory itself.

`xdgTestHelper` has been moved into `Tests/Tools/XDGTestHelper`.

This restructure allows for the emission of the `TestFoundation` binary
in the top level directory for the build output.  This is required to
co-locate the libraries and the executables which is needed for Windows.