Restructure `FoundationXML` into `Sources/FoundationXML`
This splits out the `FoundationXML` sources from `Foundation` into
`Sources/FoundationXML`.  This makes it easier to identify the target
that the sources are associated with and to grep for something in a
particular target.

Restructure the build to follow best practices for CMake and split out
the monolithic CMakeLists.txt into a target specific one.